Move all mod integrations to 'compat' package
