+++ src/main/kotlin/com/kneelawk/wiredredstone/compat/emi/SimpleRedstoneAssemblerEmiRecipe.kt
@@ -1,4 +1,4 @@
-package com.kneelawk.wiredredstone.emi
+package com.kneelawk.wiredredstone.compat.emi
 
 import com.kneelawk.wiredredstone.WRConstants.gui
 import com.kneelawk.wiredredstone.WRConstants.tooltip
@@ -7,7 +7,7 @@ import com.kneelawk.wiredredstone.blockentity.RedstoneAssemblerBlockEntity.Compa
 import com.kneelawk.wiredredstone.blockentity.RedstoneAssemblerBlockEntity.Companion.CRAFTING_SLOT_COUNT
 import com.kneelawk.wiredredstone.blockentity.RedstoneAssemblerBlockEntity.Companion.OUTPUT_HEIGHT
 import com.kneelawk.wiredredstone.blockentity.RedstoneAssemblerBlockEntity.Companion.OUTPUT_WIDTH
-import com.kneelawk.wiredredstone.emi.WiredRedstonePlugin.Companion.SPRITE_SHEET
+import com.kneelawk.wiredredstone.compat.emi.WiredRedstonePlugin.Companion.SPRITE_SHEET
 import com.kneelawk.wiredredstone.recipe.RedstoneAssemblerRecipe
 import com.kneelawk.wiredredstone.recipe.RedstoneAssemblerShapedRecipe
 import com.kneelawk.wiredredstone.recipe.RedstoneAssemblerShapelessRecipe
