+++ src/main/kotlin/com/kneelawk/wiredredstone/compat/rei/WiredRedstoneCommonPlugin.kt
@@ -1,4 +1,4 @@
-package com.kneelawk.wiredredstone.rei
+package com.kneelawk.wiredredstone.compat.rei
 
 import com.kneelawk.wiredredstone.screenhandler.RedstoneAssemblerScreenHandler
 import me.shedaniel.rei.api.common.display.DisplaySerializerRegistry
