+++ src/main/kotlin/com/kneelawk/wiredredstone/compat/wthit/PartPlugin.kt
@@ -1,4 +1,4 @@
-package com.kneelawk.wiredredstone.wthit
+package com.kneelawk.wiredredstone.compat.wthit
 
 import alexiil.mc.lib.multipart.api.AbstractPart
 import alexiil.mc.lib.multipart.api.MultipartUtil
